tools/tooling/requirements.txt
jsonschema
pynacl
cryptography
requests
